refactor: move font & theme onto assets folder
