+++ client/src/App.tsx
@@ -1,25 +1,12 @@
-import '@fontsource/roboto/300.css';
-import '@fontsource/roboto/400.css';
-import '@fontsource/roboto/500.css';
-import '@fontsource/roboto/700.css';
-
-import { red } from '@mui/material/colors';
-import CssBaseline from '@mui/material/CssBaseline';
-import { createTheme, ThemeProvider } from '@mui/material/styles';
-import { Snackbars } from 'components/Snackbar';
-import { UsernameDialoug } from 'components/UsernameDialoug';
+import { Route, Routes } from 'react-router';
+import { CssBaseline, ThemeProvider } from 'components/MaterialUI';
 import { Home } from 'features/Home';
 import { RoomPage } from 'features/Room';
+import { Snackbars } from 'components/Snackbar';
 import { SnackbarProvider } from 'notistack';
-import { Route, Routes } from 'react-router';
-
-const theme = createTheme({
-  palette: {
-    primary: {
-      main: red[500],
-    },
-  },
-});
+import { UsernameDialoug } from 'components/UsernameDialoug';
+import { theme } from 'assets/theme';
+import 'assets/fonts';
 
 export default function App() {
   return (
